Separate work scheduler from BoC
The BoC runtime is built ontop of a work stealing scheduler.  The work stealing scheduler does not require the BoC runtime.  The restructures the code to make the work stealing scheduler more standalone.
