+++ src/rt/sched/work.h
@@ -2,10 +2,9 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
-#include "epoch.h"
-
 #include <atomic>
 #include <snmalloc/snmalloc.h>
+#include "../ds/heap.h"
 
 namespace verona::rt
 {
